+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 * Provides a new ANS top-level class/schema to enable Story objects to refer to Lists<ANS> instead of just an HTML String blob
 * Implements new types of ANS content like Text, Blockquote, etc
 * Sub-classed empty Story and Gallery class from what *used* to be called Story and is now called Collection.
+* Added a "version" property at the root-level ANS object to hold the value of the ANS schema the ANS object conforms to
 
 ## 0.2.0 2015/08/27